test: add search component test
